+++ emacs-macros/src/lib.rs
@@ -4,6 +4,7 @@
 #![recursion_limit = "128"]
 
 extern crate proc_macro;
+extern crate proc_macro2;
 
 use proc_macro::TokenStream;